1268. Search Suggestions System Trie

+++ src/main/java/com/leetcode/algo/string/search_suggest_sys_1268/SearchTrie.java
@@ -14,7 +14,7 @@ public class SearchTrie {
     }
 
     private static int getRadix(char c) {
-        return (int)(c - 'a');
+        return c - 'a';
     }
 
     private final Node root = new Node();
